提交 修改 扫描数量改为小数

+++ 3DHomeProject/KEEPER.K3.3D.Core/ParamOption/ConvertOption.cs
@@ -57,7 +57,7 @@ namespace KEEPER.K3._3D.Core.ParamOption
         /// <summary>
         /// 合格或不合格数量
         /// </summary>
-        public List<int> mount { get; set; }
+        public List<decimal> mount { get; set; }
         /// <summary>
         /// 源单行号
         /// </summary>
